recomendation_service/pyproject.toml
[tool.poetry]
name = "recomendation-service"
version = "0.1.0"
description = ""
authors = ["squeakbug <squeakbug73@outlook.com>"]

packages = [
    { include = "src" },
]

[tool.poetry.dependencies]
python = "^3.11"
requests = {git = "https://github.com/psf/requests"}
pymorphy3 = "^1.3.1"
python-telegram-bot = "^20.7"
pymorphy3-dicts-ru = "^2.4.417150.4580142"
python-dotenv = "^1.0.0"
nltk = "^3.8.1"
dataclass-csv = "^1.4.0"
numpy = "^1.26.2"
seaborn = "^0.13.0"
networkx = "^3.2.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
black = "^23.12.1"
isort = "^5.13.2"
pre-commit = "^3.6.0"

[tool.flake8]
max-line-length = 99

[tool.black]
line-length = 99
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
force_grid_wrap = 2
atomic = true
include_trailing_comma = true
lines_after_imports = 2
lines_between_types = 1
use_parentheses = true
src_paths = ["src",]
filter_files = true
known_first_party = ["src"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
